Updates to run in C++98 standard.

+++ include/ShiftInfo.h
@@ -19,7 +19,8 @@
 
 namespace prp  {
 
-typedef uint32_t ulongint;
+// typedef uint32_t ulongint;
+typedef unsigned long ulongint;
 
 class ShiftInfo {
 	public:
